Pull out regex into a utility
