+++ src/util/types.ts
@@ -1,0 +1,6 @@
+// regex to identify a bytes type
+export const bytesPattern = /^bytes\d*$/;
+// regex to identify digits
+export const digitsPattern = /\d+$/;
+// regex to identify a uint type
+export const uintPattern = /^uint\d+$/;
